# api/agent/swagger.yaml
basePath: /v1
definitions:
  api.appSessionResponse:
    properties:
      app_id:
        type: string
      issued_at:
        type: string
      token:
        type: string
      workflows_run:
        items:
          type: string
        type: array
    type: object
  api.appSummary:
    properties:
      id:
        type: string
      version:
        type: string
      workflows_run:
        items:
          type: string
        type: array
    type: object
  api.connectorSecretRef:
    properties:
      key:
        type: string
      type:
        type: string
    type: object
  api.connectorSummary:
    properties:
      config:
        additionalProperties: {}
        type: object
      created_at:
        type: string
      id:
        type: string
      secret_refs:
        additionalProperties:
          $ref: '#/definitions/api.connectorSecretRef'
        type: object
      type:
        type: string
    type: object
  api.connectorTestResponse:
    properties:
      message:
        type: string
      ok:
        type: boolean
    type: object
  api.createConnectorRequest:
    properties:
      config:
        additionalProperties: {}
        type: object
      id:
        type: string
      secret_refs:
        additionalProperties:
          $ref: '#/definitions/api.connectorSecretRef'
        type: object
      type:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    Public HTTP API for the Patchcord agent (vision document, section 10.1). Covers workflow triggering and run observability, connector CRUD and testing, application hosting, and a read-only plugin catalog listing; the rest of the vision document's API surface (actions, full plugin management) is not implemented yet.
    Every route marked with a lock icon below requires "Authorization: Bearer <admin token>" — but only once at least one admin token has been created (`patchcord auth token create`); a fresh agent answers every request unauthenticated, exactly as before this existed (ADR-0036).
  title: Patchcord Agent API
  version: "1"
paths:
  /apps:
    get:
      description: Returns every installed application and the permissions its sessions
        are limited to.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.appSummary'
            type: array
      security:
      - BearerAuth: []
      summary: List installed applications
      tags:
      - apps
  /apps/{id}/sessions:
    post:
      description: 'Issues a new session for the named application, limited to its
        manifest''s declared permissions. The token is a bearer credential: pass it
        as "Authorization: Bearer <token>" to POST /workflows/{id}/run.'
      parameters:
      - description: App id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.appSessionResponse'
        "404":
          description: app not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Issue an application session
      tags:
      - apps
  /connectors:
    get:
      description: Returns every recorded connector, ordered by id. Never includes
        a resolved secret value — only each secret reference's logical type/key.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.connectorSummary'
            type: array
      security:
      - BearerAuth: []
      summary: List connectors
      tags:
      - connectors
    post:
      consumes:
      - application/json
      description: 'Creates a new connector: a persistent, named configuration for
        accessing an external system. type must match a connector type declared by
        an installed plugin''s manifest.'
      parameters:
      - description: Connector to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.createConnectorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.connectorSummary'
        "400":
          description: malformed request body, or the connector is invalid (empty
            id/type, unknown type, unsupported secret reference type)
          schema:
            type: string
        "409":
          description: a connector with this id already exists
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a connector
      tags:
      - connectors
  /connectors/{id}:
    delete:
      description: Removes a connector. There is no update endpoint (ADR-0020) — a
        caller that wants to change one recreates it, which is exactly what the dashboard's
        "Modifier" does under the hood.
      parameters:
      - description: Connector id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: deleted
        "404":
          description: connector not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a connector
      tags:
      - connectors
    get:
      description: Returns one connector's configuration and secret references (never
        a resolved secret value).
      parameters:
      - description: Connector id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.connectorSummary'
        "404":
          description: connector not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a connector
      tags:
      - connectors
  /connectors/{id}/test:
    post:
      description: Resolves the connector's configuration and secrets, then asks the
        installed plugin that declares its type to actually attempt a connection.
        Unlike GET /connectors/{id}, this calls into a live plugin process. A connection
        attempt that runs but fails (e.g. wrong password) is reported as ok=false,
        not an HTTP error.
      parameters:
      - description: Connector id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.connectorTestResponse'
        "404":
          description: connector not found
          schema:
            type: string
        "500":
          description: no connector tester configured, a secret failed to resolve,
            or the plugin call itself failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Test a connector
      tags:
      - connectors
...
  /runs/{id}/events:
    get:
      description: Streams a run's status changes, and its steps', as Server-Sent
        Events until the run reaches a terminal status or the client disconnects.
        The "event:" field is "run.<status>" or "step.<status>"; "data:" is a JSON-encoded
        event payload.
      parameters:
      - description: Run id
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: text/event-stream
          schema:
            type: string
        "404":
          description: run not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Stream a run's events
      tags:
      - runs
...
securityDefinitions:
  BearerAuth:
    description: Admin token, if any has been created (see the top-level description).
      Pass as "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"